requirements.txt
streamlit==1.28.0
pdfplumber==0.9.0
pandas==2.0.3
openai==0.28.1
python-dotenv==1.0.0
latex2mathml==3.77.0
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.0.1
openpyxl==3.1.2
tabulate==0.9.0
PyMuPDF==1.23.8
xlsxwriter==3.1.9
python-docx==0.8.11
streamlit-cropper==0.0.3
matplotlib==3.7.2